src/test/java/com/rideshare/PopupTest.java
import java.util.List;

public class PopupTest {

    public static void main(String[] args) {
        testPopupCreation();
        testPopupOpenClose();
        testPopupButtonClick();
    }

    public static void testPopupCreation() {
        Popup popup = new Popup("Confirmation", "Are you sure you want to delete this item?", List.of("Yes", "No"),
                "Open", "Close", "Cancel");

        assert popup.getTitle().equals("Confirmation");
        assert popup.getText().equals("Are you sure you want to delete this item?");
        assert popup.getButtons().equals(List.of("Yes", "No"));
        assert popup.getConfirmButtonText().equals("Open");
        assert popup.getCancelButtonText().equals("Close");
        assert popup.getCloseButtonText().equals("Cancel");
    }

    public static void testPopupOpenClose() {
        Popup popup = new Popup("Test", "Test popup", List.of("OK"), "Open", "Close", "Cancel");

        popup.open();
        // Here you can add assertions to check if the popup is open

        popup.close();
        // Here you can add assertions to check if the popup is closed
    }

    public static void testPopupButtonClick() {
        Popup popup = new Popup("Test", "Test popup", List.of("OK"), "Open", "Close", "Cancel");

        popup.onSubmit("OK");
        // Here you can add assertions to verify that the correct action is taken when a
        // button is clicked
    }
}
